SSSIHL #69: Sum elements at indices with exactly k set bits.
